Create CODE-2.1
This code uses the digitalRead() function to read the sensor values and the analogWrite() function to control the speed and direction of the robot's motors. The control algorithm simply checks the sensor readings and moves the robot forward if the left and right sensors detect the line, turns the robot in the direction of the front sensors if they detect the line, and stops the motors otherwise.

You will need to modify this code to match the specific hardware and sensors you are using, as well as to implement any additional features or behaviors you want your line-following robot to have.